// template/react-starter-kit/src/.generated/api/model/createTermGroupRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateTermGroupRequestDto {
  /** @maxLength 50 */
  code: string;
  /** @maxLength 255 */
  title: string;
  isRequired?: boolean;
  sortOrder?: number;
}